Receive Data Messages on iOS (#16)
When the app is in the foreground, data message will be delivered to Flutterland via `onMessage` directly. If the app is not in the foreground, messages will be stored and delivered to the app via `onMessage` when the app is brought to the foreground again. This is a limitation of Firebase (see https://firebase.google.com/docs/cloud-messaging/concept-options#data_messages).